Move `users` subcommand funcs from user.go to users.go
